requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
geode-background==9.2.4
    # via
    #   geode-explicit
    #   geode-implicit
    #   geode-simplex
geode-common==33.8.0
    # via
    #   -r requirements.in
    #   geode-background
    #   geode-conversion
    #   geode-explicit
    #   geode-implicit
    #   geode-numerics
    #   geode-simplex
    #   geode-viewables
geode-conversion==6.2.12
    # via
    #   geode-explicit
    #   geode-implicit
geode-explicit==6.1.39
    # via
    #   -r requirements.in
    #   geode-implicit
geode-implicit==3.7.10
    # via -r requirements.in
geode-numerics==6.0.6
    # via
    #   -r requirements.in
    #   geode-implicit
    #   geode-simplex
geode-simplex==9.3.1
    # via
    #   -r requirements.in
    #   geode-implicit
geode-viewables==3.1.1
    # via -r requirements.in
opengeode-core==15.18.0
    # via
    #   -r requirements.in
    #   geode-background
    #   geode-common
    #   geode-conversion
    #   geode-explicit
    #   geode-implicit
    #   geode-numerics
    #   geode-simplex
    #   geode-viewables
    #   opengeode-geosciences
    #   opengeode-geosciencesio
    #   opengeode-inspector
    #   opengeode-io
opengeode-geosciences==8.4.6
    # via
    #   -r requirements.in
    #   geode-implicit
    #   geode-viewables
    #   opengeode-geosciencesio
opengeode-geosciencesio==5.3.9
    # via
    #   -r requirements.in
    #   geode-implicit
opengeode-inspector==6.4.1
    # via
    #   -r requirements.in
    #   geode-explicit
    #   geode-implicit
    #   geode-simplex
opengeode-io==7.2.0
    # via
    #   -r requirements.in
    #   geode-implicit
    #   geode-viewables
    #   opengeode-geosciencesio
